* [html5] bugfix: the unit wx doesn't work correctly.
Since the wx unit was designed to match the native unit px, so the '1wx'
should be the same long in any devices or any screens. It doesn't adapt
to screens by multiply a ratio which is `weex.config.env.scale`. Just
keep them in px forms and be their original values.